+++ apps/server/createHaiyama.ts
@@ -1,4 +1,4 @@
-import { type Hai, constructHai } from "../src/utils/hai.js";
+import { type Hai, constructHai } from "shared/hai.ts";
 
 function shuffleArray<T>(array: T[]): T[] {
 	for (let i = array.length - 1; i > 0; i--) {
