+++ apps/shared/package.json
@@ -1,0 +1,10 @@
+{
+	"name": "shared",
+	"private": true,
+	"devDependencies": {
+		"@types/bun": "latest"
+	},
+	"peerDependencies": {
+		"typescript": "^5"
+	}
+}
