+++ apps/web/src/utils/calculateSyantenMentsu.ts
@@ -1,4 +1,4 @@
-import type { Hai, HaiKind, JihaiValue } from "./hai.ts";
+import type { Hai, HaiKind, JihaiValue } from "shared/hai";
 
 /** 抜き出した面子と面子候補の数、及び対子を含むかの情報 */
 export type MentsuCountInfo = {
