+++ apps/web/vite.config.ts
@@ -4,5 +4,6 @@ import { defineConfig } from "vite";
 
 // https://vitejs.dev/config/
 export default defineConfig({
+	envDir: "../../",
 	plugins: [react(), tailwindcss()],
 });
