+++ package.json
@@ -3,42 +3,20 @@
 	"private": true,
 	"version": "0.0.0",
 	"type": "module",
+	"workspaces": ["apps/*"],
 	"scripts": {
-		"dev": "concurrently \"bun vite\" \"bun run server/server.ts\"",
+		"dev": "concurrently \"cd apps/web && bun dev\" \"cd apps/server && bun dev\"",
 		"build": "bun vite build",
 		"lint": "eslint .",
 		"preview": "bun vite preview",
 		"format": "bunx biome check . --write"
 	},
-	"dependencies": {
-		"@emotion/react": "^11.13.3",
-		"@emotion/styled": "^11.13.0",
-		"@mui/base": "^5.0.0-beta.58",
-		"@mui/icons-material": "^6.1.3",
-		"@mui/material": "^6.0.0",
-		"@tailwindcss/vite": "^4.1.7",
-		"concurrently": "^9.1.0",
-		"material-react-table": "^3.0.1",
-		"react": "^18.3.1",
-		"react-dom": "^18.3.1",
-		"react-icons": "^5.3.0",
-		"react-markdown": "^9.0.1",
-		"react-router-dom": "^6.26.2",
-		"tailwindcss": "^4.1.7"
-	},
+	"dependencies": {},
 	"devDependencies": {
 		"@biomejs/biome": "^1.9.4",
-		"@eslint/js": "^9.9.0",
 		"@types/bun": "^1.2.10",
-		"@types/css-modules": "^1.0.5",
-		"@types/react": "^18.3.3",
-		"@types/react-dom": "^18.3.0",
-		"@vitejs/plugin-react": "^4.3.1",
-		"css-loader": "^7.1.2",
+		"concurrently": "^9.1.0",
 		"lefthook": "^1.11.13",
-		"style-loader": "^4.0.0",
-		"typescript": "^5.5.3",
-		"typescript-eslint": "^8.0.1",
-		"vite": "^5.4.10"
+		"typescript": "^5.5.3"
 	}
 }
