+++ shell.nix
@@ -5,10 +5,10 @@
   prisma =
     (nix-prisma-utils.lib.prisma-factory {
       inherit pkgs;
-      prisma-fmt-hash = "sha256-atD5GZfmeU86mF1V6flAshxg4fFR2ews7EwaJWZZzbc=";
-      query-engine-hash = "sha256-8FTZaKmQCf9lrDQvkF5yWPeZ7TSVfFjTbjdbWWEHgq4=";
-      libquery-engine-hash = "sha256-USIdaum87ekGY6F6DaL/tKH0BAZvHBDK7zjmCLo//kM=";
-      schema-engine-hash = "sha256-k5MkxXViEqojbkkcW/4iBFNdfhb9PlMEF1M2dyhfOok=";
+      prisma-fmt-hash = "sha256-ROsmQvRXtpClfj/de8hifTc4FVCMNT7u2Qwie+G7l1Y=";
+      query-engine-hash = "sha256-bIkXzxjR7exW1US2XJAFedpCo7huuDjDIUE4bGmSSs0=";
+      libquery-engine-hash = "sha256-8VL8/jmWR325PXFwrzIoNSTtRxiQ9SXDjXoUmfeVxgU=";
+      schema-engine-hash = "sha256-jSM/yfKACWAFwmbXDDL9VO1oGIiILyYDFXXTfcSWbwA=";
     }).fromBunLock
     ./bun.lock;
 in
